+++ lib/address_search_field.dart
@@ -2,8 +2,8 @@
 /// It can be used to get Directions beetwen two points.
 library address_search_field;
 
-export 'package:address_search_field/src/widgets/address_search_field.dart';
-export 'package:address_search_field/src/widgets/address_search_dialog.dart';
+export 'package:address_search_field/src/widgets/address_field.dart';
+export 'package:address_search_field/src/widgets/address_dialog.dart';
 export 'package:address_search_field/src/widgets/route_search_box.dart';
 export 'package:address_search_field/src/services/geo_methods.dart';
 export 'package:address_search_field/src/models/directions.dart';